requirements.txt
Django==1.8.5
Markdown==2.6.3
apuestas==0.1
argparse==1.2.1
django-filter==0.11.0
djangorestframework==3.3.0
wheel==0.26.0
wsgiref==0.1.2
